+++ CHANGELOG.md
@@ -294,3 +294,4 @@
 * SlideshowControllerEx now tries to use dialogue characters and backgrounds as slides
 * OnFire and OnReload scripts are now passed item model in addition to instance
 * Added PlaySound overloads with more params (like PlaySoundEx, but does not throw)
+* Added options to not receive messages on inactive components (QdmsMessageInterface and QdmsMessageComponent)